requirements.txt
# Основные зависимости
fastapi==0.103.1
uvicorn==0.23.2

# Работа с Docker
docker==7.0.0

# Сбор метрик системы
psutil==5.9.6

# Анализ сетевого трафика
scapy==2.5.0

# Телеграм-бот
python-telegram-bot==20.3

# Управление настройками и валидация
pydantic-settings==2.0.3
pydantic==2.5.1

# Асинхронная работа с файлами
aiofiles==23.2.1

# Дополнительные утилиты
python-dotenv==1.0.0  # Для загрузки переменных окружения из .env
requests==2.31.0
